Move ranking evaluation rake task into its own file
