+++ lib/tasks/debug.rake
@@ -2,7 +2,6 @@ require "aws-sdk-s3"
 require "csv"
 require "rummager"
 require "rainbow"
-require "debug/rank_eval"
 require "debug/synonyms"
 require "tempfile"
 
@@ -56,36 +55,6 @@ namespace :debug do
         puts title
         puts description if description
         puts ""
-      end
-    end
-  end
-
-  desc "Check how well the search query performs for a set of relevancy judgements"
-  task :ranking_evaluation, [:datafile, :ab_tests] do |_, args|
-    csv = args.datafile || begin
-      bucket = ENV["AWS_S3_RELEVANCY_BUCKET_NAME"]
-      raise "Missing required AWS_S3_RELEVANCY_BUCKET_NAME envvar" if bucket.nil?
-
-      csv = Tempfile.open(["judgements", ".csv"])
-      Services.s3_client.get_object(bucket:, key: "judgements.csv", response_target: csv.path)
-      csv.path
-    end
-
-    begin
-      evaluator = Debug::RankEval.new(csv, args.ab_tests)
-      results = evaluator.evaluate
-
-      maxlen = results[:query_scores].map { |query, _| query.length }.max
-      results[:query_scores].each do |query, score|
-        justified_query = "#{query}:".ljust(maxlen + 1)
-        puts "#{justified_query} #{score}"
-      end
-      puts "---"
-      puts "overall score: #{results[:score]}"
-    ensure
-      if csv.is_a?(Tempfile)
-        csv.close
-        csv.unlink
       end
     end
   end
